bug fixed in countries filter-sort mechanism and also add map's search bar ui is finished

+++ app/src/main/java/com/wenubey/countryapp/data/local/entities/CountryFavEntity.kt
@@ -13,7 +13,7 @@ data class CountryFavEntity(
     val countryOfficialName: String?,
     val countryNativeName: Map<String, NativeNameEntity>?,
     val capital: List<String>?,
-    val population: Double?,
+    val population: Int?,
     val topLevelDomain: List<String>?,
     val countryCodeCCA2: String?,
     val isIndependent: Boolean?,
